Renamed java.py to java_analysis.py for PEP8 compliance
