+++ tests/analysis/java/test_java_analysis.py
